[LI-CHERRY-PICK] KAFKA-13630: Reduce amount of time that producer network thread holds batch queue lock (#11722) (#391)
TICKET = LIKAFKA-46436

LI_DESCRIPTION =
This cherry picks from apache#11722
Hold the deque lock for only as long as is required to collect and make a decision in
ready() and drain() loops. Once this is done, remaining work can be done without lock,
so release it. This allows producers to continue appending.

EXIT_CRITERIA = N/A